pyproject.toml
[tool.poetry]
name = "dodi-assessment"
version = "0.1.0"
description = ""
authors = ["oscaroguledo <122152636+oscaroguledo@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
celery = "^5.3.6"
Django = "^5.0.3"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
jsonfield = "^3.1.0"
pillow = "^10.2.0"
pytest = "^8.1.1"

[tool.poetry.group.dev.dependencies]
djangorestframework = "^3.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
